Fix bug
After adding the threshold slider, everytime the slider value changes,
another API query needs to be run.
The names had to be saved to be used when the slider changed.
The initial value was set to an empty vector, when in reality it should
be set to an empty string.